Python/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
backoff==1.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
googleapis-common-protos==1.56.4
grpcio==1.47.0
idna==3.3
opentelemetry-api==1.12.0rc2
opentelemetry-distro==0.32b0
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-instrumentation==0.32b0
opentelemetry-instrumentation-requests==0.32b0
opentelemetry-proto==1.12.0rc2
opentelemetry-sdk==1.12.0rc2
opentelemetry-semantic-conventions==0.32b0
opentelemetry-util-http==0.32b0
protobuf==3.20.1
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1